% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_and_export_means}
\alias{calculate_and_export_means}
\title{Calculate Mean Values of Specified Columns and Export to CSV}
\usage{
calculate_and_export_means(data, out_dir, file_name = "table_mean.csv")
}
\arguments{
\item{data}{A dataframe or a tibble containing the relevant columns.}

\item{out_dir}{A string specifying the directory path where the output CSV
file will be saved.}

\item{file_name}{A string specifying the name of the output CSV file.
Default is "table_mean.csv".}
}
\value{
Writes a CSV file to the specified directory containing the mean
values for the specified range of columns. The function itself
invisibly returns NULL.
}
\description{
Calculate mean values for specified columns in a dataframe and exports
these mean values to a CSV file. It is specifically designed to process
columns related to socioeconomic indicators, from 'Income_2021' to 'CWB_2021'
The function is designed to assist in the preprocessing steps of a data
analysis workflow.
}
\details{
The function operates on a range of columns, calculating the mean
for each column and exporting the results to a CSV file named
"table_mean.csv" by default. It's part of a larger analysis workflow
process that aims to learn about the health
and happiness of communities.
}
\examples{
df <- tibble::tibble(
  Income_2021 = c(50000, 60000, 70000),
  Education_2021 = c(12, 14, 16),
  Housing_2021 = c(200, 250, 300),
  Labour_Force_Activity_2021 = c(1, 2, 1),
  CWB_2021 = c(70, 75, 80)
)

out_dir <- "results"
if (!dir.exists(out_dir)) {
dir.create(out_dir)}
calculate_and_export_means(df, out_dir, "table_mean.csv")
if (dir.exists(out_dir)) {
unlink(out_dir, recursive = TRUE)
}

}
