% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fetch_data}
\alias{fetch_data}
\title{Reads data from a given URl string literal and outputs that data}
\usage{
fetch_data(data_url, tfHeader, sepType)
}
\arguments{
\item{data_url}{A URl String to the .csv file online}

\item{tfHeader}{Boolean TRUE or FALSE for if the csv file has a header row}

\item{sepType}{String for how the values in the csv file has been separated}
}
\value{
A data frame or data frame extension (e.g. a tibble).
}
\description{
This function takes a URl to a csv file online, whether the csv file has an header
and the separation type as an input
and returns that data.
}
\examples{
data_url <- paste0("https://data.sac-isc.gc.ca/",
            "geomatics/directories/output/",
            "DonneesOuvertes_OpenData/CWB/",
            "CWB_2021.csv")
fetch_data(data_url, TRUE, ",")
# Specify the URL for the data
}
